% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScores.R
\name{getScores}
\alias{getScores}
\title{Title Get GPAs and GRE scores from a SOPHAS full application PDF}
\usage{
getScores(pdfdir)
}
\arguments{
\item{pdfdir}{A full-specified path to a directory containing one or more
SOPHAS application PDFs}
}
\value{
a data.frame with applicants in rows and data in columns
}
\description{
Title Get GPAs and GRE scores from a SOPHAS full application PDF
}
